+++ pakyow-core/lib/pakyow/generator/helpers.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Pakyow
+  class Generator
+    module Helpers
+      def dot
+        "."
+      end
+    end
+  end
+end