+++ pakyow-core/spec/unit/generator_spec.rb
@@ -7,34 +7,42 @@ RSpec.describe Pakyow::Generator do
     File.expand_path("../support/generator", __FILE__)
   end
 
+  let :destination_path do
+    File.expand_path("../support/generated", __FILE__)
+  end
+
+  let :generator_class do
+    Class.new(described_class)
+  end
+
+  let :instance do
+    generator_class.new(source_path, **options)
+  end
+
+  let(:options) {
+    {}
+  }
+
   describe "#initialize" do
     it "initializes with a path" do
       expect {
-        described_class.new(source_path)
+        generator_class.new(source_path)
       }.to_not raise_error
     end
   end
 
   describe "#generate" do
-    let :destination_path do
-      File.expand_path("../support/generated", __FILE__)
-    end
-
     let :options do
       {
         foo: "bar"
       }
     end
 
-    let :instance do
-      described_class.new(source_path)
-    end
-
     it "generates each file" do
       expect(instance.files.length).to eq(5)
 
       instance.files.each do |file|
-        expect(file).to receive(:generate).with(destination_path, options)
+        expect(file).to receive(:generate).with(Pathname.new(destination_path), context: instance_of(generator_class))
       end
 
       instance.generate(destination_path, options)
@@ -44,11 +52,11 @@ RSpec.describe Pakyow::Generator do
       before do
         $generate_hook_calls = []
 
-        described_class.before "generate" do
+        generator_class.before "generate" do
           $generate_hook_calls << :before
         end
 
-        described_class.after "generate" do
+        generator_class.after "generate" do
           $generate_hook_calls << :after
         end
 
@@ -61,72 +69,83 @@ RSpec.describe Pakyow::Generator do
         expect($generate_hook_calls[1]).to be(:after)
       end
     end
+
+    describe "actions" do
+      before do
+        local = self
+
+        generator_class.action do
+          local.calls << :action
+        end
+
+        allow_any_instance_of(Pakyow::Generator::File).to receive(:generate)
+      end
+
+      let(:calls) {
+        []
+      }
+
+      it "calls actions" do
+        instance.generate(destination_path, options)
+
+        expect(calls).to eq([:action])
+      end
+    end
   end
 
   describe "#run" do
-    let :instance do
-      described_class.new(source_path)
-    end
-
     let :runner_double do
       instance_double(Pakyow::Support::CLI::Runner)
     end
 
-    it "runs the command in context of the destination" do
-      instance.instance_variable_set(:@destination_path, "dest")
+    before do
+      generator_class.action do
+        run "ls", message: "foo"
+      end
 
+      allow_any_instance_of(Pakyow::Generator::File).to receive(:generate)
+    end
+
+    it "runs the command in context of the destination" do
       expect(Pakyow::Support::CLI::Runner).to receive(:new).with(message: "foo").and_return(
         runner_double
       )
 
-      expect(runner_double).to receive(:run).with("cd dest && ls")
+      expect(runner_double).to receive(:run).with("cd #{destination_path} && ls")
 
-      instance.run("ls", message: "foo")
+      instance.generate(destination_path)
     end
-  end
-end
 
-RSpec.describe Pakyow::Generator::File do
-  let :source_path do
-    File.expand_path("../support/generator", __FILE__)
-  end
+    context "run is called out of context of perform" do
+      it "runs the command in context of the current directory" do
+        expect(Pakyow::Support::CLI::Runner).to receive(:new).with(message: "foo").and_return(
+          runner_double
+        )
 
-  let :file_path do
-    "test.txt"
-  end
+        expect(runner_double).to receive(:run).with("cd . && ls")
 
-  let :destination_path do
-    File.expand_path("../support/generated", __FILE__)
-  end
-
-  let :instance do
-    generator_class.new(File.join(source_path, file_path), source_path)
-  end
-
-  let :generator_class do
-    described_class
-  end
-
-  describe "#initialize" do
-    it "initializes with a path and source path" do
-      expect {
-        described_class.new(File.join(source_path, file_path), source_path)
-      }.to_not raise_error
+        instance.run("ls", message: "foo")
+      end
     end
   end
 
-  describe "#generate" do
+  describe "generated files" do
+    let(:source_path) {
+      File.expand_path("../support/generator/test.txt", __FILE__)
+    }
+
+    let(:generated_file_path) {
+      "test.txt"
+    }
+
     before do
       cleanup
-      instance.generate(destination_path, options)
+
+      instance.generate(destination_path)
     end
 
     after do
       cleanup
-    end
-
-    let :options do
-      {}
     end
 
     def cleanup
@@ -136,26 +155,30 @@ RSpec.describe Pakyow::Generator::File do
     end
 
     it "generates the file at the destination" do
-      expect(File.exist?(File.join(destination_path, file_path))).to be(true)
+      expect(File.exist?(File.join(destination_path, generated_file_path))).to be(true)
     end
 
     describe "the generated file" do
       it "contains the correct content" do
-        expect(File.read(File.join(destination_path, file_path))).to eq("this is a plain file test\n")
+        expect(File.read(File.join(destination_path, generated_file_path))).to eq("this is a plain file test\n")
       end
     end
 
     context "file is an erb template" do
-      let :file_path do
-        "test.txt.erb"
-      end
+      let(:source_path) {
+        File.expand_path("../support/generator/test.txt.erb", __FILE__)
+      }
+
+      let(:generated_file_path) {
+        "test.txt"
+      }
 
       it "renames the generated file" do
-        expect(File.exist?(File.join(destination_path, "test.txt"))).to be(true)
+        expect(File.exist?(File.join(destination_path, generated_file_path))).to be(true)
       end
 
       it "evaluates the template" do
-        expect(File.read(File.join(destination_path, "test.txt")).strip).to eq("2")
+        expect(File.read(File.join(destination_path, generated_file_path)).strip).to eq("2")
       end
 
       context "options are passed" do
@@ -173,39 +196,47 @@ RSpec.describe Pakyow::Generator::File do
     end
 
     context "file is a dotfile" do
-      let :file_path do
-        "%dot%gitignore"
-      end
+      let(:source_path) {
+        File.expand_path("../support/generator/%dot%gitignore", __FILE__)
+      }
+
+      let(:generated_file_path) {
+        ".gitignore"
+      }
 
       it "renames the generated file" do
-        expect(File.exist?(File.join(destination_path, ".gitignore"))).to be(true)
+        expect(File.exist?(File.join(destination_path, generated_file_path))).to be(true)
       end
 
       describe "the generated file" do
         it "contains the correct content" do
-          expect(File.read(File.join(destination_path, ".gitignore")).strip).to eq("dotfile_test")
+          expect(File.read(File.join(destination_path, generated_file_path)).strip).to eq("dotfile_test")
         end
       end
     end
 
     context "file is a keep file" do
-      let :file_path do
-        "test-keep/keep"
-      end
+      let(:source_path) {
+        File.expand_path("../support/generator/test-keep", __FILE__)
+      }
 
       it "generates the enclosing directory" do
-        expect(File.directory?(File.join(destination_path, "test-keep"))).to be(true)
+        expect(File.directory?(File.join(destination_path, "folder"))).to be(true)
       end
 
       it "does not generate the keep file" do
-        expect(File.exist?(File.join(destination_path, "test-keep/keep"))).to be(false)
+        expect(File.exist?(File.join(destination_path, "folder/keep"))).to be(false)
       end
     end
 
     context "with a subclass" do
-      let :file_path do
-        "test-methods.txt.erb"
-      end
+      let(:source_path) {
+        File.expand_path("../support/generator/test-methods.txt.erb", __FILE__)
+      }
+
+      let(:generated_file_path) {
+        "test-methods.txt"
+      }
 
       let :generator_class do
         Class.new(described_class) do
@@ -216,7 +247,7 @@ RSpec.describe Pakyow::Generator::File do
       end
 
       it "exposes content through methods" do
-        expect(File.read(File.join(destination_path, "test-methods.txt")).strip).to eq("bar")
+        expect(File.read(File.join(destination_path, generated_file_path)).strip).to eq("bar")
       end
     end
   end